requirements.txt
reportlab~=4.2.2
pytest~=8.2.2
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pytest-xdist~=3.6.1
pyyaml~=6.0.1

